AFIRGEN FINAL/asr ocr model server/requirements.txt
fastapi==0.116.1
Pillow==11.3.0
pydantic==2.11.7
torch==2.7.1
transformers==4.52.4
uvicorn==0.35.0
whisper==1.1.10
filelock==3.16.1 
ffmpeg-python==0.2.0 